Bugfix | V1-239 - UI changes
