+++ frontend/src/components/Layout/AuthorizedLayout/AuthorizedLayoutContainer.tsx
@@ -1,5 +1,4 @@
 import { FC, useState } from 'react';
-import { useLocation } from 'react-router-dom';
 
 import { useLogOut } from 'api/logOut';
 import { useGetUserInfo } from 'api/userInfo';
@@ -13,15 +12,13 @@ import AuthorizedLayout from './AuthorizedLayout';
 const EMPTY_ARGUMENT = null;
 
 const AuthorizedLayoutContainer: FC = () => {
-  const { pathname } = useLocation();
-
   const [isMobileMenuOpen, toggleMobileMenuOpen] = useToggle();
   const [isSqueeze, toggleSqueeze] = useToggle();
   const [isConfirmOpen, setConfirmOpen] = useState<boolean>(false);
 
   const { data: profileInfoData, isLoading: isProfileLoading } = useGetUserInfo();
 
-  const { mutate: logOutMutate, isLoading: isLogOutLoading } = useLogOut(pathname);
+  const { mutate: logOutMutate, isLoading: isLogOutLoading } = useLogOut();
 
   const { avatar, firstName, lastName, notifications } = profileInfoData || {};
   const userInfo = { avatar, firstName, lastName };
