+++ frontend/src/components/Layout/AuthorizedLayout/Menu/Badge/utils.ts
@@ -1,7 +1,7 @@
 import { BadgeType } from 'enums/badgeType';
 import { getUserProfileCache } from 'utils/cache';
 
-export const getBadgeContentByType = (type?: BadgeType) => {
+export const getBadgeContentByType = (type?: BadgeType): number | undefined => {
   const userProfileCache = getUserProfileCache();
 
   const USER_FIELDS_BY_TYPE = {
