+++ frontend/src/components/SkillGroup/SkillItem/index.ts
@@ -1,1 +1,1 @@
-export { default } from './SkillItemContainer';
+export { default } from './SkillItem';
