+++ frontend/src/components/Tooltip/Tooltip.tsx
@@ -1,8 +1,14 @@
 import { FC } from 'react';
-import { Tooltip as MuiTooltip, TooltipProps as ITooltipProps } from '@mui/material';
+import { Tooltip as MuiTooltip, TooltipProps as ITooltipProps, Fade } from '@mui/material';
 
 const Tooltip: FC<ITooltipProps> = ({ children, ...props }) => (
-  <MuiTooltip placement="top-start" {...props}>
+  <MuiTooltip
+    placement="top-start"
+    enterTouchDelay={0}
+    TransitionComponent={Fade}
+    TransitionProps={{ timeout: 0 }}
+    {...props}
+  >
     {children}
   </MuiTooltip>
 );
