+++ frontend/src/pages/PendingAssessments/RequestItem/RequestTechnologies/TechnologyItem/TechnologyItem.tsx
@@ -1,27 +1,25 @@
 import { FC } from 'react';
 
 import Avatar from 'components/Avatar';
+import Tooltip from 'components/Tooltip';
 import { Size } from 'enums/sizes';
 import { ITechnologyItemProps } from 'pages/PendingAssessments/types';
 
-import { HoverSkillInfoText, TechnologyWrapper } from './styled';
+import { TechnologyWrapper } from './styled';
 
-const TechnologyItem: FC<ITechnologyItemProps> = ({
-  technology,
-  isInfoShown,
-  toggleFullInfoShown,
-}) => {
+const TechnologyItem: FC<ITechnologyItemProps> = ({ technology }) => {
   const { skill, points: technologyPoints } = technology || {};
   const { name: skillName, maxScore: skillMaxScore, image: skillImage } = skill || {};
 
+  const skillScoreText = `${skillName}:
+  (${technologyPoints}/${skillMaxScore})`;
+
   return (
-    <TechnologyWrapper onMouseEnter={toggleFullInfoShown} onMouseLeave={toggleFullInfoShown}>
-      <Avatar size={Size.xxsmall} avatar={skillImage} />
-      {isInfoShown && (
-        <HoverSkillInfoText>{`${skillName}:
-      (${technologyPoints}/${skillMaxScore})`}</HoverSkillInfoText>
-      )}
-    </TechnologyWrapper>
+    <Tooltip title={skillScoreText}>
+      <TechnologyWrapper>
+        <Avatar size={Size.xxsmall} avatar={skillImage} />
+      </TechnologyWrapper>
+    </Tooltip>
   );
 };
 export default TechnologyItem;
