+++ frontend/src/pages/PendingAssessments/RequestItem/RequestTechnologies/TechnologyItem/index.ts
@@ -1,1 +1,1 @@
-export { default } from './TechnologyItemContainer';
+export { default } from './TechnologyItem';
