+++ frontend/src/pages/Profile/UserSkills/SkillInfo/index.ts
@@ -1,0 +1,1 @@
+export { default } from './SkillInfo';
