+++ frontend/src/pages/Profile/types.ts
@@ -1,0 +1,9 @@
+import { IUserSkill } from 'types/skill';
+
+export interface ISkillInfoProps {
+  skillItem: IUserSkill;
+}
+
+export interface ISkillInfoWrapperProps {
+  completed?: boolean;
+}
