+++ frontend/src/pages/Profile/utils/calculateSkillProgress.ts
@@ -1,0 +1,4 @@
+const PERCENTAGE = 100;
+
+export const calculateSkillProgress = (stagesCompleted: number, maxStages: number): number =>
+  (stagesCompleted / maxStages) * PERCENTAGE;
