+++ frontend/src/pages/Profile/utils/index.ts
@@ -1,0 +1,1 @@
+export { calculateSkillProgress } from './calculateSkillProgress';
